bluesky_grokker/requirements.txt
atproto>=0.0.46
websockets>=12.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
redis>=5.0.1
numpy>=1.26.0
openai>=1.12.0
anthropic>=0.39.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
httpx>=0.27.0
modal>=0.64.0
